+++ src/components/Item/VisualizationItem/styles/ItemHeaderButtons.module.css
@@ -1,0 +1,7 @@
+.itemActionsWrap {
+    margin-left: auto;
+    flex-shrink: 0;
+}
+.itemActionsWrap button {
+    margin-left: var(--spacers-dp8);
+}
