+++ src/components/ItemGrid/ItemGrid.css
@@ -48,26 +48,6 @@
     background: none;
 }
 
-/* dashboard item - header */
-
-.dashboard-item-header {
-    font-size: 15px;
-    font-weight: 700;
-    color: #3a3a3a;
-    display: flex;
-    align-items: center;
-    padding: 0 4px 0 14px;
-    height: 45px;
-}
-
-.dashboard-item-header-title {
-    flex: 1;
-    overflow: hidden;
-    position: relative;
-    top: -1px;
-    left: -1px;
-}
-
 /* dashboard item - content */
 
 .dashboard-item-content {
